# docs/swagger.json
{"swagger":"2.0","info":{"description":"This service supports Create, Read, Update, Delete of the Employee model","version":"1.0.0","title":"Employee CRUD Service","termsOfService":"http://swagger.io/terms/","contact":{"email":"apiteam@swagger.io"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:8002","basePath":"/api","tags":[{"name":"Employee","description":"Employee CRUD Service"},{"name":"employee"}],"schemes":["http"],"paths":{"/employee/publish":{"post":{"tags":["employee"],"summary":"Publish new employee details to Kafka","description":"","operationId":"publish","consumes":["application/json","application/xml"],"produces":["application/json","application/xml"],"parameters":[{"in":"body","name":"body","description":"Employee Details to add","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"405":{"description":"Invalid input"}}}},"/employee/consume":{"get":{"tags":["employee"],"summary":"Consume from Kafka and write to MongoDB","description":"","operationId":"consume","produces":["application/json","application/xml"],"parameters":[],"responses":{"405":{"description":"Invalid input"}}}},"/employee":{"post":{"tags":["employee"],"summary":"Add new Employee","description":"","operationId":"createEmployee","consumes":["application/json","application/xml"],"produces":["application/json","application/xml"],"parameters":[{"in":"body","name":"body","description":"Employee Details to add","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"405":{"description":"Invalid input"}}},"put":{"tags":["employee"],"summary":"Update an existing employee","description":"","operationId":"updateEmployee","consumes":["application/json","application/xml"],"produces":["application/json","application/xml"],"parameters":[{"in":"body","name":"body","description":"Employee to update","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"400":{"description":"Invalid ID supplied"},"404":{"description":"Employee not found"},"405":{"description":"Validation exception"}}},"patch":{"tags":["employee"],"summary":"Patch an existing employee","description":"","operationId":"patchEmployee","consumes":["application/json","application/xml"],"produces":["application/json","application/xml"],"parameters":[{"in":"body","name":"body","description":"Employee to update","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"400":{"description":"Invalid ID supplied"},"404":{"description":"Employee not found"},"405":{"description":"Validation exception"}}}},"/employee/{employeeId}":{"get":{"tags":["employee"],"summary":"Find Employee by ID","description":"Returns employee information with id","operationId":"getEmployeeById","produces":["application/json","application/xml"],"parameters":[{"name":"employeeId","in":"path","description":"ID of Employee to return","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Employee"}},"400":{"description":"Invalid ID supplied"},"404":{"description":"Employee not found"}}},"delete":{"tags":["employee"],"summary":"Deletes a employee","description":"","operationId":"deleteEmployee","produces":["application/json","application/xml"],"parameters":[{"name":"employeeId","in":"path","description":"Employee id to delete","required":true,"type":"integer","format":"int32"}],"responses":{"400":{"description":"Invalid ID supplied"},"404":{"description":"Employee not found"}}}}},"definitions":{"Employee":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"designation":{"type":"string"}},"xml":{"name":"Employee"}}},"externalDocs":{"description":"Find out more about Swagger","url":"http://swagger.io"}}